OtherEvents - more organization - remove PaperEvents since this plugin requires paper now
